move entity deletion from Entities to AllStorages move Key to its own file

+++ src/storage/all/view.rs
@@ -1,4 +1,4 @@
-use super::{Key, Storage, TypeIdHasher};
+use super::{Entities, Key, Storage, TypeIdHasher};
 use std::any::TypeId;
 use std::collections::HashMap;
 use std::hash::BuildHasherDefault;
@@ -10,24 +10,61 @@ pub struct AllStoragesViewMut<'a>(
 );
 
 impl AllStoragesViewMut<'_> {
-    pub(crate) fn delete(&mut self, entity: Key) {
-        let mut storage_to_unpack = Vec::new();
-        for storage in self.0.values_mut() {
-            let observers = storage.delete(entity).unwrap();
-            storage_to_unpack.reserve(observers.len());
-            let mut i = 0;
-            for observer in observers.iter().copied() {
-                while i < storage_to_unpack.len() && observer < storage_to_unpack[i] {
-                    i += 1;
-                }
-                if storage_to_unpack.is_empty() || observer != storage_to_unpack[i] {
-                    storage_to_unpack.insert(i, observer);
+    /// Delete an entity and all its components.
+    /// Returns `true` if `entity` was alive.
+    /// # Example
+    /// ```
+    /// # use shipyard::prelude::*;
+    /// let world = World::new::<(usize, u32)>();
+    ///
+    /// let mut entity1 = None;
+    /// let mut entity2 = None;
+    /// world.run::<(EntitiesMut, &mut usize, &mut u32), _>(|(mut entities, mut usizes, mut u32s)| {
+    ///     entity1 = Some(entities.add_entity((&mut usizes, &mut u32s), (0usize, 1u32)));
+    ///     entity2 = Some(entities.add_entity((&mut usizes, &mut u32s), (2usize, 3u32)));
+    /// });
+    ///
+    /// world.run::<AllStorages, _>(|mut all_storages| {
+    ///     all_storages.delete(entity1.unwrap());
+    /// });
+    ///
+    /// world.run::<(&usize, &u32), _>(|(usizes, u32s)| {
+    ///     assert_eq!((&usizes).get(entity1.unwrap()), None);
+    ///     assert_eq!((&u32s).get(entity1.unwrap()), None);
+    ///     assert_eq!(usizes.get(entity2.unwrap()), Some(&2));
+    ///     assert_eq!(u32s.get(entity2.unwrap()), Some(&3));
+    /// });
+    /// ```
+    pub fn delete(&mut self, entity: Key) -> bool {
+        let mut entities = self.0[&TypeId::of::<Entities>()].entities_mut().unwrap();
+
+        if entities.delete_key(entity) {
+            drop(entities);
+
+            let mut storage_to_unpack = Vec::new();
+
+            for storage in self.0.values_mut() {
+                let observers = storage.delete(entity).unwrap();
+                storage_to_unpack.reserve(observers.len());
+
+                let mut i = 0;
+                for observer in observers.iter().copied() {
+                    while i < storage_to_unpack.len() && observer < storage_to_unpack[i] {
+                        i += 1;
+                    }
+                    if storage_to_unpack.is_empty() || observer != storage_to_unpack[i] {
+                        storage_to_unpack.insert(i, observer);
+                    }
                 }
             }
-        }
 
-        for storage in storage_to_unpack {
-            self.0.get_mut(&storage).unwrap().unpack(entity).unwrap();
+            for storage in storage_to_unpack {
+                self.0.get_mut(&storage).unwrap().unpack(entity).unwrap();
+            }
+
+            true
+        } else {
+            false
         }
     }
 }
